Homework

+++ optimizations/JPEG/Images/Matrix.cs
@@ -7,16 +7,16 @@ namespace JPEG.Images
         public readonly Pixel[,] Pixels;
         public readonly int Height;
         public readonly int Width;
-				
+
         public Matrix(int height, int width)
         {
             Height = height;
             Width = width;
-			
-            Pixels = new Pixel[height,width];
-            for(var i = 0; i< height; ++i)
-            for(var j = 0; j< width; ++j)
-                Pixels[i, j] = new Pixel(0, 0, 0, PixelFormat.RGB);
+
+            Pixels = new Pixel[height, width];
+            for (var i = 0; i < height; ++i)
+                for (var j = 0; j < width; ++j)
+                    Pixels[i, j] = new Pixel(0, 0, 0, PixelFormat.RGB);
         }
 
         public static explicit operator Matrix(Bitmap bmp)
@@ -25,9 +25,9 @@ namespace JPEG.Images
             var width = bmp.Width - bmp.Width % 8;
             var matrix = new Matrix(height, width);
 
-            for(var j = 0; j < height; j++)
+            for (var j = 0; j < height; j++)
             {
-                for(var i = 0; i < width; i++)
+                for (var i = 0; i < width; i++)
                 {
                     var pixel = bmp.GetPixel(i, j);
                     matrix.Pixels[j, i] = new Pixel(pixel.R, pixel.G, pixel.B, PixelFormat.RGB);
@@ -39,11 +39,11 @@ namespace JPEG.Images
 
         public static explicit operator Bitmap(Matrix matrix)
         {
-            var bmp = new Bitmap(matrix.Width, matrix.Height);
+            var bmp = new Bitmap((int)matrix.Width, (int)matrix.Height);
 
-            for(var j = 0; j < bmp.Height; j++)
+            for (var j = 0; j < bmp.Height; j++)
             {
-                for(var i = 0; i < bmp.Width; i++)
+                for (var i = 0; i < bmp.Width; i++)
                 {
                     var pixel = matrix.Pixels[j, i];
                     bmp.SetPixel(i, j, Color.FromArgb(ToByte(pixel.R), ToByte(pixel.G), ToByte(pixel.B)));
@@ -55,7 +55,7 @@ namespace JPEG.Images
 
         public static int ToByte(double d)
         {
-            var val = (int) d;
+            var val = (int)d;
             if (val > byte.MaxValue)
                 return byte.MaxValue;
             if (val < byte.MinValue)
